# openapi.yaml
openapi: 3.0.1
info:
  title: Transparent TikTok API
  version: 0.1.0
  description: |
    Core endpoints for user, video, interaction management,
    plus feed and explanation APIs.
servers:
  - url: https://api.transparent-tiktok.com/v1
paths:
  /users:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{user_id}:
    get:
      summary: Get user profile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    UserCreate:
      type: object
      required: [username, email_hash]
      properties:
        username:
          type: string
        display_name:
          type: string
        email_hash:
          type: string
    User:
      allOf:
        - $ref: '#/components/schemas/UserCreate'
        - type: object
          properties:
            user_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            last_active_at:
              type: string
              format: date-time
...